requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
fuzzywuzzy==0.18.0
idna==3.4
Levenshtein==0.20.9
python-magic==0.4.27
rapidfuzz==2.13.7
requests==2.28.2
titlecase==2.4
urllib3==1.26.14
